% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMET-package.R
\docType{package}
\name{scMET-package}
\alias{scMET-package}
\alias{.datatable.aware}
\title{\code{scMET}: Bayesian modelling of DNA methylation at single-cell
resolution.}
\format{
An object of class \code{logical} of length 1.
}
\usage{
.datatable.aware
}
\value{
scMET main package documentation.
}
\description{
Package for analysing single-cell DNA methylation datasets.
scMET performs feature selection, by identifying highly variable features,
and also differential testing, based on mean but also more importantly on
variability between two groups of cells.
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R
package version 2.19.3. https://mc-stan.org
}
\seealso{
\code{\link{scmet}}, \code{\link{scmet_differential}},
\code{\link{scmet_hvf}}
}
\author{
C.A.Kapourani \email{kapouranis.andreas@gmail.com}
}
\keyword{datasets}
