% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMET-package.R
\docType{data}
\name{scmet_diff_dt}
\alias{scmet_diff_dt}
\title{Synthetic methylation data from two groups of cells}
\format{
An object of class \code{scmet_simulate_diff} of length 9.
}
\usage{
scmet_diff_dt
}
\value{
A list object with simulated data.
}
\description{
Small synthetic data for quick analysis, mostly useful for
showing the differential analysis one can perform using scMET.
}
\keyword{datasets}
