% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMET-package.R
\docType{data}
\name{scmet_dt}
\alias{scmet_dt}
\title{Synthetic methylation data from a single population}
\format{
An object of class \code{scmet_simulate} of length 5.
}
\usage{
scmet_dt
}
\value{
A list object with simulated data.
}
\description{
Small synthetic data for quick analysis, mostly useful for
performing feature selection and capturing mean-variance relationship with
scMET.
}
\keyword{datasets}
