% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_design_matrix}
\alias{create_design_matrix}
\title{Create design matrix}
\usage{
create_design_matrix(L, X, c = 1.2)
}
\arguments{
\item{L}{Total number of basis functions, including the bias term.}

\item{X}{Vector of covariates}

\item{c}{Scaling parameter for variance of RBFs}
}
\value{
A design matrix object H.
}
\description{
Generic function for crating a radial basis function (RBF)
design matrix for input vector X.
}
\seealso{
\code{\link{scmet}}, \code{\link{scmet_differential}},
\code{\link{scmet_hvf_lvf}}
}
\author{
C.A.Kapourani \email{C.A.Kapourani@ed.ac.uk}
}
